% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_continuous.R
\name{fit_continuous}
\alias{fit_continuous}
\title{Train models for continuous label}
\usage{
fit_continuous(data, frac = 0.8, models = c("LM", "SVM"))
}
\arguments{
\item{data}{A data frame with feature and label.}

\item{frac}{The proportion of train_data. The default value is 0.8.}

\item{models}{A vector of model names want to be trained. The default value is c("LM","SVM").}
}
\value{
A list of MSE and models.
}
\description{
This function can automatically split the train test datasets, and train LM and SVM models for continuous label and gain the MSE on test set and models. Also plot the prediction and ground truth of test data.
}
\examples{
\dontrun{
set.seed(516)
x1 <- 1:100
x2 <- seq(2, 18, length.out=100) + rnorm(100, 3, 4)
y <- 4x1 - 3x2 + rnorm(100, 0, 2)
data <- data.frame(x1=x1, x2=x2, y=y)
results <- fit_continuous(data, frac=0.8, models=c("LM","SVM"))
}
}
