% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAIDP.R
\name{SAIDP}
\alias{SAIDP}
\title{Survival analysis in digital pathology}
\usage{
SAIDP(data, data2, cutpoint)
}
\arguments{
\item{data}{Basic information of each patient in the dataset, such as survival time, survival state, age, etc.}

\item{data2}{Information about TIL of each patient obtained by neural network model.}

\item{cutpoint}{Determine the size of the TIL Score based on the cutpoint.}
}
\value{
A series of survival analyses of datasets.
}
\description{
This function can conduct data preprocessing for TIL related information obtained by the model and obtain a series of survival analysis results.
}
\examples{
\dontrun{
#load("data\\data.rda")
#load("data\\data2.rda")
#SAIDP(data=data, data2=data2, cutpoint=0.0002820585)
}
}
