\name{StatComp22017-package}
\alias{StatComp22017-package}
\docType{package}
\title{
  Functions in R and Rcpp for digital pathology and 'Statistical Computing' course.
}
\description{
  Several R and Rcpp functions for digital pathology and statistical computing are included in the package.
}
\author{
Ziqiao Zhou

Maintainer: Ziqiao Zhou <zzq2001@mail.ustc.edu.cn>
}
\keyword{ package }
