% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{til}
\alias{til}
\title{Compute TIL Score using Rcpp}
\usage{
til(LYM, STR, TUM, TIL3, TIL7, TIL8)
}
\arguments{
\item{LYM}{the number of patches of LYM}

\item{STR}{the number of patches of STR}

\item{TUM}{the number of patches of TUM}

\item{TIL3}{the TIL Score of LYM}

\item{TIL7}{the TIL Score of STR}

\item{TIL8}{the TIL Score of TUM}
}
\description{
Compute TIL Score using Rcpp
}
